Move common instrumentation tests utils to libtesting-ui
