+++ instrumentation-tests/src/androidTest/java/com/mapbox/navigation/instrumentation_tests/core/BannerAndVoiceInstructionsTest.kt
@@ -6,14 +6,14 @@ import com.mapbox.navigation.base.options.HistoryRecorderOptions
 import com.mapbox.navigation.base.options.NavigationOptions
 import com.mapbox.navigation.core.MapboxNavigationProvider
 import com.mapbox.navigation.instrumentation_tests.activity.EmptyTestActivity
-import com.mapbox.navigation.instrumentation_tests.utils.coroutines.bannerInstructions
-import com.mapbox.navigation.instrumentation_tests.utils.coroutines.sdkTest
-import com.mapbox.navigation.instrumentation_tests.utils.coroutines.voiceInstructions
-import com.mapbox.navigation.instrumentation_tests.utils.coroutines.withLogOnTimeout
 import com.mapbox.navigation.instrumentation_tests.utils.history.MapboxHistoryTestRule
 import com.mapbox.navigation.instrumentation_tests.utils.routes.RoutesProvider
 import com.mapbox.navigation.instrumentation_tests.utils.routes.RoutesProvider.toNavigationRoutes
 import com.mapbox.navigation.testing.ui.BaseTest
+import com.mapbox.navigation.testing.ui.utils.coroutines.bannerInstructions
+import com.mapbox.navigation.testing.ui.utils.coroutines.sdkTest
+import com.mapbox.navigation.testing.ui.utils.coroutines.voiceInstructions
+import com.mapbox.navigation.testing.ui.utils.coroutines.withLogOnTimeout
 import com.mapbox.navigation.testing.ui.utils.getMapboxAccessTokenFromResources
 import junit.framework.Assert.assertEquals
 import kotlinx.coroutines.async
