+++ libtesting-ui/src/main/java/com/mapbox/navigation/testing/ui/utils/MapboxNavigationRule.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.instrumentation_tests.utils
+package com.mapbox.navigation.testing.ui.utils
 
 import com.mapbox.navigation.core.MapboxNavigationProvider
 import com.mapbox.navigation.testing.ui.utils.runOnMainSync
